[WOLFPAL-52] Updated landing page to remove marketing sections.
