--- client/components/Landing/CTA.vue
@@ -1,26 +1,0 @@
-<template>
-  <div class="bg-gray-200 dark:bg-gray-900">
-    <div
-      class="px-4 py-12 mx-auto max-w-7xl sm:px-6 lg:py-16 lg:px-8 lg:flex lg:items-center lg:justify-between"
-    >
-      <h2
-        class="text-3xl font-extrabold tracking-tight text-gray-900 dark:text-white sm:text-4xl"
-      >
-        <span class="block">Ready to complete your goals?</span>
-        <span class="block text-primary-500 dark:text-primary-400"
-          >Start your free trial today.</span
-        >
-      </h2>
-      <div class="flex mt-8 lg:mt-0 lg:flex-shrink-0">
-        <div class="inline-flex rounded-md shadow">
-          <nuxt-link
-            to=""
-            class="inline-flex items-center justify-center px-5 py-3 text-base font-medium text-white border border-transparent rounded-md bg-primary-500 hover:bg-primary-600"
-          >
-            Get started
-          </nuxt-link>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
