+++ client/layouts/landing.vue
@@ -1,0 +1,7 @@
+<template>
+  <div>
+    <portal to="destination"></portal>
+    <SharedOfflineBanner v-if="$nuxt.isOffline" />
+    <Nuxt />
+  </div>
+</template>
